backend/requirements/development.txt
# Include base requirements
-r base.txt

# Development & Debugging
django-debug-toolbar>=4.3,<5.0
ipython>=8.0,<9.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=5.0,<6.0
factory-boy>=3.3,<4.0

# Code Quality
black>=24.0,<25.0
ruff>=0.4,<1.0
isort>=5.13,<6.0
mypy>=1.10,<2.0
django-stubs>=5.0,<6.0

# Pre-commit hooks
pre-commit>=3.7,<4.0










